+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Added:
+- Phone field with country selector
+
 ## [2.0.8] - 2024-07-09
 ### Added:
 - Flow to redeposit 2FA expired coins